add loading for hide/show/delete of expert, show expert sections with… (#370)
* add loading for hide/show/delete of expert, show expert sections with no data when editing

* update theme package for quicklinks profile icon

+++ services/base-service/spa/client/public/elements/pages/expert/app-expert.js
@@ -77,6 +77,8 @@ export default class AppExpert extends Mixin(LitElement)
    * @return {Object} e
    */
   async _onAppStateUpdate(e) {
+    this.expertImpersonating = utils.getCookie('impersonateId');
+
     if( e.location.page !== 'expert' ) return;
     window.scrollTo(0, 0);
 
@@ -392,10 +394,13 @@ export default class AppExpert extends Mixin(LitElement)
     this.showModal = false;
 
     if( this.isAdmin && this.modalAction === 'hide-expert' ) {
+      this.dispatchEvent(new CustomEvent("loading", {}));
       try {
         let res = await this.ExpertModel.updateExpertVisibility(this.expertId, false);
+        this.dispatchEvent(new CustomEvent("loaded", {}));
         this.isVisible = false;
       } catch (error) {
+        this.dispatchEvent(new CustomEvent("loaded", {}));
         let modelContent = `<p>Hiding expert could not be done through Aggie Experts right now. Please, try again later, or make changes directly in the <a href="https://oapolicy.universityofcalifornia.edu/">UC Publication Management System.</a></p>`;
 
         this.modalTitle = 'Error: Update Failed';
@@ -407,13 +412,15 @@ export default class AppExpert extends Mixin(LitElement)
         this.hideOaPolicyLink = true;
         this.errorMode = true;
       }
-    } else if( this.modalAction === 'delete-expert' ) {
+    } else if( this.isAdmin && this.modalAction === 'delete-expert' ) {
+      this.dispatchEvent(new CustomEvent("loading", {}));
       try {
         let res = await this.ExpertModel.deleteExpert(this.expertId);
-
+        this.dispatchEvent(new CustomEvent("loaded", {}));
         // redirect to home page
         this.AppStateModel.setLocation('/');
       } catch (error) {
+        this.dispatchEvent(new CustomEvent("loaded", {}));
         let modelContent = `<p>Deleting expert could not be done through Aggie Experts right now. Please, try again later, or make changes directly in the <a href="https://oapolicy.universityofcalifornia.edu/">UC Publication Management System.</a></p>`;
 
         this.modalTitle = 'Error: Update Failed';
@@ -439,10 +446,13 @@ export default class AppExpert extends Mixin(LitElement)
    */
   async _showExpert(e) {
     if( this.isAdmin ) {
+      this.dispatchEvent(new CustomEvent("loading", {}));
       try {
         let res = await this.ExpertModel.updateExpertVisibility(this.expertId, true);
+        this.dispatchEvent(new CustomEvent("loaded", {}));
         this.isVisible = true;
       } catch (error) {
+        this.dispatchEvent(new CustomEvent("loaded", {}));
         let modelContent = `<p>Showing expert could not be done through Aggie Experts right now. Please, try again later, or make changes directly in the <a href="https://oapolicy.universityofcalifornia.edu/">UC Publication Management System.</a></p>`;
 
         this.modalTitle = 'Error: Update Failed';
@@ -480,7 +490,7 @@ export default class AppExpert extends Mixin(LitElement)
   _deleteExpert(e) {
     this.modalAction = 'delete-expert';
     this.modalTitle = 'Delete Expert';
-    this.modalContent = `<p>Expert will be removed from Aggie Experts. CDL privacy will be set to private. To show the expert again in Aggie Experts, you would need to update the privacy setting to public in CDL. Are you sure you would like to continue?</p>`;
+    this.modalContent = `<p>Expert will be removed from Aggie Experts. CDL privacy will be set to "internal". To show the expert again in Aggie Experts, you would need to update the privacy setting to public in CDL. Are you sure you would like to continue?</p>`;
     this.showModal = true;
     this.hideCancel = true;
     this.hideSave = false;
